refactor: extract NewTaskForm on new archive to not trigger graphql
